Refactored dockerdeploy to use common internal repr.
We previously had a bunch of different representations of cluster
state and meta-data internally.  This refactors everything to use
a common set of types instead.
